# openapi.yaml
openapi: 3.0.3
info:
  title: School Information System API
  description: RESTful API for managing student information and interactions within a school, focused on teacher functionalities.
  version: 1.0.0
servers:
  - url: https://api.schoolsystem.com/v1
    description: Primary API endpoint

paths:
  /students:
    get:
      summary: List all students
      responses:
        '200':
          description: A list of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      summary: Enroll a new student
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewStudent'
      responses:
        '201':
          description: Student enrolled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'

  /students/{studentId}:
    get:
      summary: Retrieve a student by ID
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Details of a specific student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'

    put:
      summary: Update a student's information
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Student updated successfully
    delete:
      summary: Remove a student from enrollment
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Student removed successfully

  /teachers:
    get:
      summary: List all teachers
      responses:
        '200':
          description: A list of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'

  /courses/sections/{sectionId}/attendance:
    post:
      summary: Submit attendance for a section
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceRecord'
      responses:
        '200':
          description: Attendance recorded successfully

  /assets:
    get:
      summary: List all assets
      responses:
        '200':
          description: A list of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asset'
    post:
      summary: Add a new asset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAsset'
      responses:
        '201':
          description: Asset added successfully

components:
  schemas:
    Student:
      type: object
      properties:
        studentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        demographics:
          type: object  # Value Object containing address, contact info etc.
        enrollments:
          type: array
          items:
            type: string
        guardianEmails:
          type: array
          items:
            type: string

    NewStudent:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        demographics:
          type: object  # As above.

    Teacher:
      type: object
      properties:
        teacherId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        sections:
          type: array
          items:
            type: string

    AttendanceRecord:
      type: object
      properties:
        attendanceRecordId:
          type: string
        studentId:
          type: string
        attendanceCode:
          type: string
        comment:
          type: string

    Asset:
      type: object
      properties:
        assetId:
          type: string
        assetType:
          type: string
        description:
          type: string
        condition:
          type: string

    NewAsset:
      type: object
      properties:
        assetType:
          type: string
        description:
          type: string
        condition:
          type: string